mainControl/requirements.txt
pyserial
time
sys
socket